refactor(env): move edge details
