+++ app/react/portainer/environments/common/EdgeKeyDisplay.tsx
